Replace floats with doubles

+++ libsimulator/include/entity/steering.h
@@ -15,7 +15,7 @@ enum steering_type
 
 struct steering_path_waypoint
 {
-	float pos[2];
+	double pos[2];
 	struct steering_path_waypoint *next;
 };
 
@@ -25,10 +25,10 @@ struct component_steer;
 
 void steering_update_system(struct entity_ctx *entities);
 
-void steering_apply(struct component_steer *steer, float current_pos[2], float *velocity);
+void steering_apply(struct component_steer *steer, double current_pos[2], double *velocity);
 
-void steering_path_add(struct component_steer *steer, float waypoint[2]);
+void steering_path_add(struct component_steer *steer, double waypoint[2]);
 BOOL steering_path_pop(struct component_steer *steer);
-void steering_path_set(struct component_steer *steer, float *waypoints, uint32_t n);
+void steering_path_set(struct component_steer *steer, double *waypoints, uint32_t n);
 
 #endif
